Allow clock control to work independently of links
